lib/shadow/, lib/: sgetsgent(): Move to separate file
